feat: Separate list and detail tests on Event module
